#133 Refactored ColumnSelector to better location.
