+++ spark-cassandra-connector/src/main/scala/com/datastax/spark/connector/streaming/CassandraStreamingRDD.scala
@@ -1,8 +1,10 @@
 package com.datastax.spark.connector.streaming
 
+import com.datastax.spark.connector.{ColumnSelector, AllColumns}
+
 import scala.reflect.ClassTag
 import org.apache.spark.streaming.StreamingContext
-import com.datastax.spark.connector.rdd.{CassandraRDD, CqlWhereClause, AllColumns, ColumnSelector}
+import com.datastax.spark.connector.rdd.{CassandraRDD, CqlWhereClause}
 import com.datastax.spark.connector.rdd.reader._
 
 /** RDD representing a Cassandra table for Spark Streaming.
